constraint_system.*, generator_system.*, transformation_system.*:   Removed these classes since they are not very useful and clutter   a module we should try to keep simple.
ppl_polyhedron.*:
  Implemented a C-style wrapper around the PPL library.

polyhedron.*:
  Reimplemented the Polyhedron class to direct remove dependency on PPL.
  The data for the class is now a Matrix<R> whose columns are the vertices. 
  Added a Polytope class which represents points as a linear system.

export_numeric.cc:
  Removed scale(...) functions and interval operations.
  These should be defined using __radd__ etc in the
  Interval and BasicSet classes.

epsfstream.h:
  Made epsfstream a non-templated class and tidied much of the code.

The main aim of these changes is to improve the Polyhedron class. Ideally, the
Ariadne Polyhedron class should support the same number types as the other 
geometric classes. However, the dual-description of PPL only works over a 
field, since division is needed to go from generators to constraints. One way 
round this problem is to provide two classes, one given by generators, the 
other by constraints, which can only be converted using approximations. The
new code also has the advantage that the PPL header files do not need to be
included into polyhedron.h, which reduces dependencies (and hopefully compile
times).

Making Parallelotope a subclass of Zonotope makes sence theoretically and works
fine in C++ (and reduced the number of geometric comparisons we need), but the 
Python wrappers seem to be having some problems.

I'm trying to code all the classes so that all compiled methods (in *.tpl 
files) are in-class. This makes it easier to keep the *.cc files up-to
date. The geometric methods are now mostly inline, out-of-class functions
which call static member functions. Please let me know what you think of
this system!

FIXME: Inheritance of Parallelotope in Python.